requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
requests
bandit
safety
python-dotenv
pytest
pytest-cov